feat(测试跟踪): 新增用例页面交互及用例多人评审
--story=1010759 --user=宋昌昌 【测试跟踪】测试跟踪交互改造 --单独分支单独环境测试 https://www.tapd.cn/55049933/s/1332049

+++ test-track/frontend/src/business/case/components/public/PublicTestCaseList.vue
@@ -1,10 +1,20 @@
 <template>
 
-  <span>
-    <ms-search
-      :condition.sync="condition"
-      @search="search">
-    </ms-search>
+  <div class="case-main-layout">
+    <div class="case-main-layout-left" style="float: left; display: inline-block">
+      <ms-table-count-bar :count-content="$t('table.all_case_content') + ' (' + page.total + ')'"></ms-table-count-bar>
+    </div>
+
+    <div class="case-main-layout-right" style="float: right; display: flex">
+      <!-- 简单搜索框 -->
+      <ms-new-ui-search :condition.sync="condition" @search="search" style="float: left" />
+
+      <!-- 高级搜索框  -->
+      <ms-table-adv-search :condition.sync="condition" @search="search" ref="advanceSearch"/>
+
+      <!-- 表头自定义显示Popover  -->
+      <ms-table-header-custom-popover :fields.sync="fields" :field-key="tableHeaderKey" @reload="reloadTable" />
+    </div>
 
     <ms-table
       v-loading="loading"
@@ -18,13 +28,17 @@
       :screen-height="screenHeight"
       :batch-operators="batchButtons"
       :remember-order="true"
+      :disable-header-config="true"
       :row-order-group-id="projectId"
       :row-order-func="editTestCaseOrder"
       :fields.sync="fields"
       :field-key="tableHeaderKey"
+      :enable-header-drag="true"
       @handlePageChange="initTableData"
       @order="initTableData"
       @filter="search"
+      @callBackSelect="callBackSelect"
+      @callBackSelectAll="callBackSelectAll"
       ref="table">
 
       <span v-for="(item, index) in fields" :key="index">
@@ -83,7 +97,7 @@
         </ms-table-column>
 
         <ms-table-column
-          prop="createUser"
+          prop="createName"
           min-width="120"
           :field="item"
           :fields-width="fieldsWidth"
@@ -129,8 +143,8 @@
           :label="$t('commons.tag')"
           min-width="80">
           <template v-slot:default="scope">
-            <ms-tag v-for="(itemName,index)  in scope.row.tags" :key="index" type="success" effect="plain"
-                    :show-tooltip="scope.row.tags.length===1&&itemName.length*12<=80"
+            <ms-single-tag v-for="(itemName,index)  in scope.row.tags" :key="index" type="success" effect="plain"
+                    :show-tooltip="scope.row.tags.length === 1 && itemName.length * 12 <= 80"
                     :content="itemName" style="margin-left: 0px; margin-right: 2px"/>
             <span/>
           </template>
@@ -159,10 +173,14 @@
 
     </ms-table>
 
-    <ms-table-pagination :change="initTableData" :current-page.sync="page.currentPage" :page-size.sync="page.pageSize"
-                         :total="page.total"/>
+    <ms-table-batch-operator-group v-if="selectCounts > 0" :batch-operators="batchButtons" :select-counts="selectCounts" @clear="clearTableSelect"/>
+
+    <home-pagination v-if="page.data.length > 0 && selectCounts == 0" :change="initTableData" :current-page.sync="page.currentPage" :page-size.sync="page.pageSize"
+                     :total="page.total" layout="total, prev, pager, next, sizes, jumper" style="margin-top: 19px"/>
 
     <test-case-preview ref="testCasePreview" :loading="rowCaseResult.loading"/>
+
+    <public-test-case-show ref="publicTestCaseShow" :condition.sync="condition" :page-size.sync="page.pageSize" :page-total.sync="page.total"/>
 
     <list-item-delete-confirm
       @handleDelete="_handleDeleteVersion"
@@ -173,12 +191,18 @@
                 @copyPublic="copyPublic"
                 ref="testBatchMove"/>
 
-  </span>
+  </div>
 
 </template>
 
 <script>
-
+import MsSingleTag from "metersphere-frontend/src/components/new-ui/MsSingleTag";
+import HomePagination from '@/business/home/components/pagination/HomePagination';
+import MsTableAdvSearch from "metersphere-frontend/src/components/new-ui/MsTableAdvSearch";
+import MsTableHeaderCustomPopover from 'metersphere-frontend/src/components/new-ui/MsTableHeaderCustomPopover'
+import MsNewUiSearch from "metersphere-frontend/src/components/new-ui/MsSearch";
+import MsTableCountBar from 'metersphere-frontend/src/components/table/MsTableCountBar';
+import MsTableBatchOperatorGroup from "metersphere-frontend/src/components/new-ui/MsTableBatchOperatorGroup";
 import MsTablePagination from 'metersphere-frontend/src/components/pagination/TablePagination';
 import {TEST_CASE_CONFIGS} from "metersphere-frontend/src/components/search/search-components";
 import {TEST_CASE_LIST} from "metersphere-frontend/src/utils/constants";
@@ -197,34 +221,28 @@ import PlanStatusTableItem from "@/business/common/tableItems/plan/PlanStatusTab
 import {getCurrentProjectID, getCurrentUserId, getCurrentWorkspaceId} from "metersphere-frontend/src/utils/token";
 import {getUUID, parseTag} from "metersphere-frontend/src/utils"
 import {hasLicense} from "metersphere-frontend/src/utils/permission"
-import MsTable from "metersphere-frontend/src/components/table/MsTable";
+import MsTable from "metersphere-frontend/src/components/new-ui/MsTable";
 import MsTableColumn from "metersphere-frontend/src/components/table/MsTableColumn";
 import MsUpdateTimeColumn from "metersphere-frontend/src/components/table/MsUpdateTimeColumn";
 import MsCreateTimeColumn from "metersphere-frontend/src/components/table/MsCreateTimeColumn";
 import TestPlanCaseStatusTableItem from "@/business/common/tableItems/TestPlanCaseStatusTableItem";
 import TestCaseReviewStatusTableItem from "@/business/common/tableItems/TestCaseReviewStatusTableItem";
-import MsSearch from "metersphere-frontend/src/components/search/MsSearch";
-
 import BatchMove from "@/business/case/components/BatchMove";
 import TestCasePreview from "@/business/case/components/TestCasePreview";
 import {
-  deletePublicTestCaseVersion,
-  editTestCaseOrder,
-  getTestCase,
-  getTestCaseStep, getTestCaseVersions,
-  testCasePublicBatchCopy,
-  testCasePublicBatchDeleteToGc,
-  testCasePublicList,
+  deletePublicTestCaseVersion, editTestCaseOrder, getTestCase,
+  getTestCaseStep, getTestCaseVersions, testCaseBatchDeleteToGc, testCasePublicBatchCopy,
+  testCasePublicBatchDeleteToGc, testCasePublicList,
 } from "@/api/testCase";
-
 import ListItemDeleteConfirm from "metersphere-frontend/src/components/ListItemDeleteConfirm";
 import {TEST_CASE_STATUS_MAP} from "@/business/constants/table-constants";
 import {mapState} from "pinia";
 import {useStore} from "@/store";
 import {getProjectMemberUserFilter} from "@/api/user";
-import {operationConfirm} from "@/business/utils/sdk-utils";
 import TypeTableItem from "@/business/common/tableItems/planview/TypeTableItem";
 import {getVersionFilters} from "@/business/utils/sdk-utils";
+import {openCaseEdit} from "@/business/case/test-case";
+import PublicTestCaseShow from "@/business/case/components/public/PublicTestCaseShow"
 
 
 export default {
@@ -245,7 +263,14 @@ export default {
     ReviewStatus,
     TestCaseReviewStatusTableItem,
     TestPlanCaseStatusTableItem,
-    MsSearch
+    MsTableBatchOperatorGroup,
+    MsTableCountBar,
+    MsNewUiSearch,
+    MsTableAdvSearch,
+    MsTableHeaderCustomPopover,
+    HomePagination,
+    PublicTestCaseShow,
+    MsSingleTag
   },
   data() {
     return {
@@ -261,38 +286,48 @@ export default {
       },
       batchButtons: [
         {
-          name: this.$t('test_track.case.batch_copy'),
+          name: this.$t('test_track.case.batch_copy_btn'),
           handleClick: this.handleBatchCopy,
           permissions: ['PROJECT_TRACK_CASE:READ+BATCH_COPY']
         },
         {
-          name: this.$t('test_track.case.batch_delete_case'),
+          name: this.$t('test_track.case.batch_delete_btn'),
           handleClick: this.handleDeleteBatchToPublic,
           permissions: ['PROJECT_TRACK_CASE:READ+BATCH_DELETE'],
-        },
+          isDelete: true
+        }
       ],
       operators: [
         {
           tip: this.$t('commons.view'), icon: "el-icon-view",
           exec: this.handleEditShow,
+          isTextButton: true,
           permissions: ['PROJECT_TRACK_CASE:READ'],
         },
         {
           tip: this.$t('commons.edit'), icon: "el-icon-edit",
           exec: this.handleEdit,
+          isTextButton: true,
           permissions: ['PROJECT_TRACK_CASE:READ+EDIT'],
           isDisable: !this.isOwner
         },
         {
-          tip: this.$t('commons.copy'), icon: "el-icon-copy-document", type: "success",
-          exec: this.handleCopyPublic,
-          permissions: ['PROJECT_TRACK_CASE:READ+COPY']
-        },
-        {
-          tip: this.$t('commons.remove'), icon: "el-icon-delete", type: "danger",
-          exec: this.handleDeleteToGc,
-          permissions: ['PROJECT_TRACK_CASE:READ+DELETE'],
-          isDisable: !this.isOwner
+          isMoreOperate: true,
+          childOperate: [
+            {
+              tip: this.$t('commons.copy'), icon: "el-icon-copy-document",
+              exec: this.handleCopyPublic,
+              permissions: ['PROJECT_TRACK_CASE:READ+COPY']
+            },
+            {
+              tip: this.$t('commons.delete'), icon: "el-icon-delete",
+              exec: this.handleDeleteToGc,
+              permissions: ['PROJECT_TRACK_CASE:READ+DELETE'],
+              isDisable: !this.isOwner,
+              isDivide: true,
+              isActive: true
+            }
+          ]
         }
       ],
       page: getPageInfo(),
@@ -301,7 +336,8 @@ export default {
       rowCase: {},
       rowCaseResult: {},
       userFilter: [],
-      versionFilters: []
+      versionFilters: [],
+      selectCounts: 0
     };
   },
   props: {
@@ -364,6 +400,9 @@ export default {
     },
   },
   methods: {
+    reloadTable() {
+      this.$refs.table.resetHeader();
+    },
     initTableData(nodeIds) {
       this.condition.nodeIds = [];
       initCondition(this.condition, this.condition.selectAll);
@@ -399,26 +438,23 @@ export default {
       this.initTableData();
       this.$emit('search');
     },
-    handleEdit(testCase) {
-      let TestCaseData = this.$router.resolve({
-        path: '/track/case/all',
-        query: {
-          redirectID: getUUID(),
-          dataType: "testCase",
-          dataSelectRange: testCase.id,
-          projectId: testCase.projectId
-        }
-      });
-      window.open(TestCaseData.href, '_blank');
+    clearTableSelect() {
+      this.$refs.table.clear();
+      this.selectCounts = 0;
+    },
+    callBackSelect(selection) {
+      this.selectCounts = this.$refs.table.selectDataCounts;
+    },
+    callBackSelectAll(selection) {
+      this.selectCounts = this.$refs.table.selectDataCounts;
+    },
+    handleEdit(testCase, type) {
+      openCaseEdit(testCase.id, type, this);
     },
     handleEditShow(testCase, column) {
-      if (column.label !== this.$t('test_track.case.case_desc')) {
-        getTestCase(testCase.id)
-          .then(response => {
-            let testCase = response.data;
-            this.$emit('testCaseEditShow', testCase);
-          });
-      }
+      this.$refs.publicTestCaseShow.open(testCase.id);
+      this.$refs.publicTestCaseShow.tableData = this.page.data;
+      this.$refs.publicTestCaseShow.currentPage = this.page.currentPage;
     },
     handleDeleteToGc(testCase) {
       getTestCaseVersions(testCase.id)
@@ -427,9 +463,18 @@ export default {
             // 删除提供列表删除和全部版本删除
             this.$refs.apiDeleteConfirm.open(testCase, this.$t('test_track.case.delete_confirm'));
           } else {
-            operationConfirm(this, this.$t('test_track.case.delete_confirm') + '\'' + testCase.name + '\'', () => {
-              this._handleDeleteVersion(testCase, false);
-            });
+            let title = this.$t('test_track.case.case_delete_confirm') + ": " + testCase.name + "?";
+            this.$confirm(this.$t('test_track.case.batch_delete_tip'), title, {
+                cancelButtonText: this.$t("commons.cancel"),
+                confirmButtonText: this.$t("commons.delete"),
+                customClass: 'custom-confirm-delete',
+                callback: action => {
+                  if (action === "confirm") {
+                    this._handleDeleteVersion(testCase, false);
+                  }
+                }
+              }
+            );
           }
         });
     },
@@ -438,7 +483,7 @@ export default {
         // 删除指定版本
         deletePublicTestCaseVersion(testCase.versionId, testCase.refId)
           .then(() => {
-            this.$success(this.$t('commons.delete_success'));
+            this.$success(this.$t('commons.delete_success'), false);
             this.$refs.apiDeleteConfirm.close();
             this.$emit("refreshPublic");
           })
@@ -448,7 +493,7 @@ export default {
         param.ids.push(testCase.id);
         testCasePublicBatchDeleteToGc(param)
           .then(() => {
-            this.$success(this.$t('commons.delete_success'));
+            this.$success(this.$t('commons.delete_success'), false);
             this.$refs.apiDeleteConfirm.close();
             this.$emit("refreshPublic");
           })
@@ -479,7 +524,7 @@ export default {
     },
     handleCopyPublic(testCase) {
       this.$refs.table.selectIds.push(testCase.id);
-      this.$refs.testBatchMove.open(this.treeNodes, this.$refs.table.selectIds, this.moduleOptions);
+      this.$refs.testBatchMove.open(false, testCase.name, this.treeNodes, this.$refs.table.selectIds, this.moduleOptions);
     },
     refresh() {
       this.$refs.table.clear();
@@ -490,18 +535,28 @@ export default {
       this.$emit('refreshAll');
     },
     handleDeleteBatchToPublic() {
-      operationConfirm(this, this.$t('test_track.case.delete_confirm'), () => {
-        let param = buildBatchParam(this, this.$refs.table.selectIds);
-        testCasePublicBatchDeleteToGc(param)
-          .then(() => {
-            this.$refs.table.clear();
-            this.$emit("refreshPublic");
-            this.$success(this.$t('commons.delete_success'));
-          });
-      });
+      let title = this.$t('test_track.case.batch_delete_confirm', [this.$refs.table.selectIds.length]);
+      this.$confirm(this.$t('test_track.case.batch_delete_tip'), title, {
+          cancelButtonText: this.$t("commons.cancel"),
+          confirmButtonText: this.$t("commons.delete"),
+          customClass: 'custom-confirm-delete',
+          callback: action => {
+            if (action === "confirm") {
+              let param = buildBatchParam(this, this.$refs.table.selectIds);
+              testCasePublicBatchDeleteToGc(param)
+                .then(() => {
+                  this.clearTableSelect();
+                  this.$emit("refreshPublic");
+                  this.$success(this.$t('commons.delete_success'), false);
+                });
+            }
+          }
+        }
+      );
     },
     handleBatchCopy() {
-      this.$refs.testBatchMove.open(this.treeNodes, this.$refs.table.selectIds, this.moduleOptions);
+      let firstSelectRow = this.$refs.table.selectRows.values().next().value;
+      this.$refs.testBatchMove.open(false, firstSelectRow.name, this.treeNodes, this.$refs.table.selectIds, this.moduleOptions);
     },
     copyPublic(param) {
       param.condition = this.condition;
@@ -512,7 +567,7 @@ export default {
       testCasePublicBatchCopy(param)
         .then(() => {
           this.loading = false;
-          this.$success(this.$t('commons.save_success'));
+          this.$success(this.$t('commons.save_success'), false);
           this.$refs.testBatchMove.close();
           this.refresh();
         });
@@ -528,15 +583,31 @@ export default {
 </script>
 
 <style scoped>
-.el-table {
-  cursor: pointer;
-}
-
 .el-tag {
   margin-left: 10px;
 }
 
-:deep(.el-table) {
-  overflow: auto;
+:deep(button.el-button.el-button--default.el-button--mini) {
+  box-sizing: border-box;
+  width: 32px;
+  height: 32px;
+  background: #FFFFFF;
+  border: 1px solid #BBBFC4;
+  border-radius: 4px;
+  flex: none;
+  order: 5;
+  align-self: center;
+  flex-grow: 0;
+  margin-left: 12px;
+}
+
+:deep(button.el-button.el-button--default.el-button--mini:hover) {
+  color: #783887;
+  border: 1px solid #783887;
+}
+
+:deep(button.el-button.el-button--default.el-button--mini:focus) {
+  color: #783887;
+  border: 1px solid #783887;
 }
 </style>
